websockets: rearchitect mock server base.
