+++ proto_fuzzer/mock_server.h
@@ -14,9 +14,11 @@
 #include <curl/curl.h>
 
 #include <cstddef>
-#include <memory>
 #include <string>
 #include <vector>
+
+#include "curl_fuzzer.pb.h"
+#include "proto_fuzzer/mock_server_base.h"
 
 namespace proto_fuzzer {
 
@@ -42,26 +44,26 @@ class MockConnection {
   int client_fd_;
 };
 
-class MockServer {
+/// @class proto_fuzzer::MockServer
+/// @brief HTTP (and other stream-oriented) in-process mock peer. Writes an
+///        optional initial response synchronously on open, then pushes queued
+///        chunks one-at-a-time as curl becomes readable. Drives its own
+///        curl_multi perform loop via DriveScenario.
+class MockServer : public MockServerBase {
  public:
   MockServer();
-  ~MockServer();
-
-  MockServer(const MockServer&) = delete;
-  MockServer& operator=(const MockServer&) = delete;
-
-  void Install(CURL* easy);
+  ~MockServer() override;
 
   void SetScript(std::string initial_response, std::vector<std::string> on_readable);
 
   void DeliverNextChunk();
   bool has_more_chunks() const;
-  MockConnection* connection();
 
-  curl_socket_t HandleOpenSocket();
+ protected:
+  curl_socket_t HandleOpenSocket() override;
+  void RunLoop(CURLM* multi, CURL* easy, const curl::fuzzer::proto::Scenario& scenario) override;
 
  private:
-  std::unique_ptr<MockConnection> connection_;
   std::string initial_response_;
   std::vector<std::string> on_readable_;
   std::size_t next_chunk_;
